+++ packages/framework/resources/views/layouts/scripts.blade.php
@@ -1,5 +1,7 @@
 {{-- The compiled Vite scripts --}}
-@if(Asset::exists('app.js'))
+@if(Vite::running())
+    {{ Vite::assets(['resources/assets/app.js']) }}
+@elseif(Asset::exists('app.js'))
     <script defer src="{{ Asset::get('app.js') }}"></script>
 @endif