+++ vite.config.js
@@ -6,8 +6,35 @@ import { defineConfig } from 'vite';
 import { resolve } from 'path';
 import tailwindcss from 'tailwindcss';
 import autoprefixer from 'autoprefixer';
+import fs from 'fs';
+import path from 'path';
 
 export default defineConfig({
+    server: {
+        port: 5173,
+        hmr: {
+            host: 'localhost',
+            port: 5173,
+        },
+        middlewareMode: false,
+    },
+    plugins: [
+        {
+            name: 'hyde-vite-server',
+            configureServer(server) {
+                server.middlewares.use((req, res, next) => {
+                    if (req.url === '/') {
+                        res.end(fs.readFileSync(
+                            path.resolve(__dirname, 'vendor/hyde/realtime-compiler/resources/vite-index-page.html'),
+                            'utf-8'
+                        ));
+                    } else {
+                        next();
+                    }
+                });
+            },
+        },
+    ],
     css: {
         postcss: {
             plugins: [